requirements.txt
langchain
langchain-huggingface
langchain-pinecone
langchain-text-splitters
langchain-community
pinecone[grpc]
python-dotenv
streamlit
sentence_transformers
lark